# optimization of U256