before
```
test u256_add ... bench:     140,001 ns/iter (+/- 9,659)
test u256_mul ... bench:   4,179,748 ns/iter (+/- 420,781)
test u256_sub ... bench:     475,245 ns/iter (+/- 30,205)
```

after
```
test u256_add ... bench:     109,593 ns/iter (+/- 4,050)
test u256_mul ... bench:     818,250 ns/iter (+/- 26,664)
test u256_sub ... bench:     109,403 ns/iter (+/- 3,300)
```